requirements.txt
aiogram>=3.0.0
sentence-transformers
faiss-cpu
ollama
rank-bm25